pyproject.toml
[project]
name = "lpil-web"
version = "0.0.1"
description = "A version of PlasTeX tailored for use with LPiL documents and the Gerby-website tool."
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "plasTeX @ git+https://github.com/diSimplex/plastex.git@merged",
    "plastex-gerby-plugin @ git+https://github.com/diSimplex/plastex-gerby-plugin.git",
    "plastex-lpil-plugin @ git+https://github.com/diSimplex/plastex-lpil-plugin.git",
]
requires-python = ">=3.10"
license = {text = "Apache2.0"}

[project.scripts]
lpilWeb = "lpilWeb.cli:cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

